syncevo-dbus-server: fixed size_t printing
On 32 bit, %lu is not a proper formatter for size_t. Casting
the value to (unsigned long) to avoid this kind of problem
without causing it on 64 bit (as %u would).

+++ src/syncevo-dbus-server.cpp
@@ -2106,7 +2106,7 @@ void Connection::process(const Caller_t &caller,
 {
     SE_LOG_DEBUG(NULL, NULL, "D-Bus client %s sends %lu bytes via connection %s, %s",
                  caller.c_str(),
-                 message.first,
+                 (unsigned long)message.first,
                  getPath(),
                  message_type.c_str());
